+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This project adheres to [Semantic Versioning](http://semver.org/). All notable c
 
 ## [Unreleased](https://github.com/OldSneerJaw/borealis-pg-cli/compare/v1.3.0...HEAD)
 - Adds the `borealis-pg:restore:capabilities` command to retrieve an add-on's database restore capabilities
+- Adds the `borealis-pg:restore:execute` command to perform an add-on database restore/clone
 
 ## [1.3.0](https://github.com/OldSneerJaw/borealis-pg-cli/compare/v1.2.0...v1.3.0)
 - Adds the `borealis-pg:integrations` command to retrieve a list of data integrations for an add-on